pyproject.toml
[project]
name = "agent-based"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=4.0.0",
    "ipykernel>=7.0.1",
    "jupyter>=1.1.1",
    "mango-agents>=2.1.5",
    "notebook>=7.4.7",
]
